requirements.txt
Django==4.2
djangorestframework==3.14.0
python-decouple==3.5
Pillow==10.0.0
stripe==5.0.0
django-cors-headers==3.14.0
django-filter==23.2.0
django-allauth==0.52.0